【frame】添加tcpclient,fs_addr优化 socket uint64类型

+++ frame/include/FrightenStone/common/basedefs/DataType/ForAll/Defs/NetDataType.h
@@ -36,7 +36,7 @@
 #ifdef _WIN32
 typedef unsigned long long MYSOCKET;
 #else
-typedef int MYSOCKET;
+typedef unsigned long long MYSOCKET;
 #endif
 
 #endif
